room saves entire chat

+++ app/src/main/java/com/example/chatapp/database/api/ChatListAPI.java
@@ -7,10 +7,12 @@ import androidx.lifecycle.MutableLiveData;
 import com.example.chatapp.database.dao.ChatDao;
 import com.example.chatapp.database.entities.Chat;
 import com.example.chatapp.database.entities.ChatDetails;
-import com.example.chatapp.database.subentities.User;
+import com.example.chatapp.database.entities.User;
+import com.example.chatapp.database.subentities.Message;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -26,6 +28,7 @@ public class ChatListAPI {
     String JWT;
     Retrofit retrofit;
     WebServiceAPI webServiceAPI;
+    User oldUser;
 
     public ChatListAPI(MutableLiveData<List<ChatDetails>> chatListData, ChatDao chatDao, String serverURL, String JWT) {
         this.chatListData = chatListData;
@@ -38,10 +41,23 @@ public class ChatListAPI {
                 .addConverterFactory(GsonConverterFactory.create(gson))
                 .build();
         webServiceAPI = retrofit.create(WebServiceAPI.class);
+        new Thread(()->oldUser=chatDao.getUser()).start();
+    }
 
-        // todo: JWT should already be in shared preferences. After implementing login,
-        //  remove this line.
-//        getToken("hello", "Helloworld1!");
+    public User getUser(String username) {
+        Call<User> call = webServiceAPI.getUser("Bearer " + JWT, username);
+        try {
+            Response<User> response = call.execute();
+            if (response.isSuccessful()) {
+                return response.body();
+            }
+            else if(response.code()==403){
+                return new User("","","unauthorized");
+            }
+            return null;
+        } catch (IOException e) {
+            return null;
+        }
     }
 
     public void getChats() {
@@ -52,9 +68,17 @@ public class ChatListAPI {
                 if (response.isSuccessful()) {
                     new Thread(() -> {
 
-                        // clears database
-                        chatDao.deleteChats();
-                        chatDao.deletePreviews();
+                        User currentUser=chatDao.getUser();
+
+                        if(currentUser!=null&&oldUser!=null&&!currentUser.getUsername().equals(oldUser.getUsername())){
+                            // clear database
+                            chatDao.deleteChats();
+                            chatDao.deletePreviews();
+                            oldUser=currentUser;
+                        }
+                        if(oldUser==null&&currentUser!=null){
+                            oldUser=currentUser;
+                        }
 
                         // converts list to array to be able to use insert(ChatDetails... chats)
                         List<ChatDetails> chatList = response.body();
@@ -64,9 +88,16 @@ public class ChatListAPI {
                         // inserts all chats (placeholder chat for the inner chats)
                         chatDao.upsert(chatArray);
                         chatDao.upsert(chatList.stream().map(cd -> {
+                            Chat oldChat=chatDao.getChat(cd.getId());
+                            if(oldChat!=null)
+                                return oldChat;
+
                             List<User> users = new ArrayList<>();
                             users.add(cd.getUser());
-                            return new Chat(cd.getId(), users, new ArrayList<>());
+                            List<Message> messageList=new ArrayList<>();
+                            if(cd.getLastMessage()!=null)
+                                messageList.add(cd.getLastMessage());
+                            return new Chat(cd.getId(), users, messageList);
                         }).toArray(Chat[]::new));
                         chatListData.postValue(chatDao.getChats());
                     }).start();
